Refactor error handling to own module
